requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
bcrypt==4.0.1
black==23.1.0
cffi==1.15.1
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
cryptography==39.0.2
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.3.1
et-xmlfile==1.1.0 ; python_version >= '3.6'
fastapi==0.93.0
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
mypy-extensions==1.0.0 ; python_version >= '3.5'
numpy==1.24.2 ; python_version >= '3.10'
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
passlib[bcrypt]==1.7.4s
pathspec==0.11.0 ; python_version >= '3.7'
platformdirs==3.1.0 ; python_version >= '3.7'
pyasn1==0.4.8
pycparser==2.21
pydantic[dotenv,email]==1.10.6
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==2.0.5.post1
starlette==0.25.0 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
uvicorn==0.20.0
